Trait orml_vesting::module::Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type Currency: LockableCurrency<Self::AccountId, Moment = Self::BlockNumber>;
    type MinVestedTransfer: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type VestedTransferOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = Self::AccountId>;
    type WeightInfo: WeightInfo;
    type MaxVestingSchedules: Get<u32>;
    type BlockNumberProvider: BlockNumberProvider<BlockNumber = Self::BlockNumber>;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type Currency: LockableCurrency<Self::AccountId, Moment = Self::BlockNumber>

type MinVestedTransfer: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

The minimum amount transferred to call vested_transfer.

type VestedTransferOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = Self::AccountId>

Required origin for vested transfer.

type WeightInfo: WeightInfo

Weight information for extrinsics in this module.

type MaxVestingSchedules: Get<u32>

The maximum vesting schedules

type BlockNumberProvider: BlockNumberProvider<BlockNumber = Self::BlockNumber>