Crate orml_vesting

Vesting Module

Overview

Vesting module provides a means of scheduled balance lock on an account. It uses the graded vesting way, which unlocks a specific amount of balance every period of time, until all balance unlocked.

Vesting Schedule

The schedule of a vesting is described by data structure VestingSchedule: from the block number of start, for every period amount of blocks, per_period amount of balance would unlocked, until number of periods period_count reached. Note in vesting schedules, time is measured by block number. All VestingSchedules under an account could be queried in chain state.

Interface

Dispatchable Functions

vested_transfer - Add a new vesting schedule for an account.

claim - Claim unlocked balances.

update_vesting_schedules - Update all vesting schedules under an account, root origin required.

Re-exports

pub use module::*;

Modules

module

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

Structs

VestingSchedule

The vesting schedule.

Constants

VESTING_LOCK_ID

Traits

WeightInfo

Weight functions needed for orml_vesting.